+++ ts/packages/knowPro/src/timestampIndex.ts
@@ -2,17 +2,9 @@
 // Licensed under the MIT License.
 
 import { collections, dateTime } from "typeagent";
-import { DateRange, IMessage, MessageIndex, TextRange } from "./dataFormat.js";
+import { DateRange, IMessage, MessageIndex } from "./dataFormat.js";
 import { textRangeFromLocation } from "./conversationIndex.js";
-
-export type TimestampedTextRange = {
-    timestamp: string;
-    range: TextRange;
-};
-
-export interface ITimestampToTextRangeIndex {
-    lookupRange(dateRange: DateRange): TimestampedTextRange[];
-}
+import { ITimestampToTextRangeIndex, TimestampedTextRange } from "./search.js";
 
 /**
  * An index of timestamp => TextRanges.
